requirements.txt
Flask~=2.0.2
requests~=2.27.1
beautifulsoup4~=4.10.0
psycopg2-binary
flask-restful
